+++ src/arrays/allUnlocks.ts
@@ -1,7 +1,4 @@
-import { CollectibleType } from "isaac-typescript-definitions";
-import { arrayRemove, copyArray } from "isaacscript-common";
 import {
-  ALT_FLOORS,
   OTHER_UNLOCK_KINDS,
   UNLOCKABLE_PATHS,
   UNLOCK_TYPES,
@@ -11,10 +8,7 @@ import type { Unlock } from "../types/Unlock";
 import { UNLOCKABLE_CARD_TYPES } from "./unlockableCardTypes";
 import { UNLOCKABLE_CHALLENGES } from "./unlockableChallenges";
 import { UNLOCKABLE_CHARACTERS } from "./unlockableCharacters";
-import {
-  BOSS_ROOM_COLLECTIBLE_TYPE_EXCEPTIONS,
-  getUnlockableCollectibleTypes,
-} from "./unlockableCollectibleTypes";
+import { UNLOCKABLE_COLLECTIBLE_TYPES } from "./unlockableCollectibleTypes";
 import { UNLOCKABLE_GRID_ENTITY_TYPES } from "./unlockableGridEntityTypes";
 import {
   UNLOCKABLE_BATTERY_SUB_TYPES,
@@ -26,14 +20,11 @@ import {
   UNLOCKABLE_SACK_SUB_TYPES,
 } from "./unlockablePickupTypes";
 import { UNLOCKABLE_PILL_EFFECTS } from "./unlockablePillEffects";
-import {
-  UNLOCKABLE_ROOM_TYPES_ONLY_NIGHTMARE,
-  getUnlockableRoomTypes,
-} from "./unlockableRoomTypes";
+import { UNLOCKABLE_ROOM_TYPES } from "./unlockableRoomTypes";
 import { UNLOCKABLE_SLOT_VARIANTS } from "./unlockableSlotVariants";
 import { UNLOCKABLE_TRINKET_TYPES } from "./unlockableTrinketTypes";
 
-const ALL_UNLOCKS: readonly Unlock[] = (() => {
+export const ALL_UNLOCKS: readonly Unlock[] = (() => {
   const unlocks: Unlock[] = [];
 
   for (const unlockType of UNLOCK_TYPES) {
@@ -62,21 +53,8 @@ const ALL_UNLOCKS: readonly Unlock[] = (() => {
         break;
       }
 
-      case UnlockType.ALT_FLOOR: {
-        for (const altFloor of ALT_FLOORS) {
-          const unlock: Unlock = {
-            type: UnlockType.ALT_FLOOR,
-            altFloor,
-          };
-          unlocks.push(unlock);
-        }
-
-        break;
-      }
-
       case UnlockType.ROOM: {
-        const unlockableRoomTypes = getUnlockableRoomTypes(false);
-        for (const roomType of unlockableRoomTypes) {
+        for (const roomType of UNLOCKABLE_ROOM_TYPES) {
           const unlock: Unlock = {
             type: UnlockType.ROOM,
             roomType,
@@ -100,8 +78,7 @@ const ALL_UNLOCKS: readonly Unlock[] = (() => {
       }
 
       case UnlockType.COLLECTIBLE: {
-        const unlockableCollectibleTypes = getUnlockableCollectibleTypes(false);
-        for (const collectibleType of unlockableCollectibleTypes) {
+        for (const collectibleType of UNLOCKABLE_COLLECTIBLE_TYPES) {
           const unlock: Unlock = {
             type: UnlockType.COLLECTIBLE,
             collectibleType,
@@ -272,36 +249,3 @@ const ALL_UNLOCKS: readonly Unlock[] = (() => {
 
   return unlocks;
 })();
-
-const ALL_UNLOCKS_NIGHTMARE: readonly Unlock[] = (() => {
-  const unlocks = copyArray(ALL_UNLOCKS);
-
-  // UnlockType.ROOM
-  for (const roomType of UNLOCKABLE_ROOM_TYPES_ONLY_NIGHTMARE) {
-    const unlock: Unlock = {
-      type: UnlockType.ROOM,
-      roomType,
-    };
-    unlocks.push(unlock);
-  }
-
-  // UnlockType.COLLECTIBLE
-  const extraNightmareCollectibleTypes = arrayRemove(
-    BOSS_ROOM_COLLECTIBLE_TYPE_EXCEPTIONS,
-    CollectibleType.BREAKFAST,
-  );
-
-  for (const collectibleType of extraNightmareCollectibleTypes) {
-    const unlock: Unlock = {
-      type: UnlockType.COLLECTIBLE,
-      collectibleType,
-    };
-    unlocks.push(unlock);
-  }
-
-  return unlocks;
-})();
-
-export function getAllUnlocks(nightmareMode: boolean): readonly Unlock[] {
-  return nightmareMode ? ALL_UNLOCKS_NIGHTMARE : ALL_UNLOCKS;
-}