+++ src/cachedEnums.ts
@@ -6,7 +6,6 @@ import {
   TrinketSlot,
 } from "isaac-typescript-definitions";
 import { getEnumValues } from "isaacscript-common";
-import { AltFloor } from "./enums/AltFloor";
 import { BossIDCustom } from "./enums/BossIDCustom";
 import { CharacterObjectiveKind } from "./enums/CharacterObjectiveKind";
 import { ObjectiveType } from "./enums/ObjectiveType";
@@ -14,8 +13,6 @@ import { OtherUnlockKind } from "./enums/OtherUnlockKind";
 import { RandomizerMode } from "./enums/RandomizerMode";
 import { UnlockType } from "./enums/UnlockType";
 import { UnlockablePath } from "./enums/UnlockablePath";
-
-export const ALT_FLOORS: readonly AltFloor[] = getEnumValues(AltFloor);
 
 export const BOSS_IDS: readonly BossID[] = getEnumValues(BossID);