+++ src/enums/UnlockType.ts
@@ -1,7 +1,6 @@
 export enum UnlockType {
   CHARACTER,
   PATH,
-  ALT_FLOOR,
   ROOM,
   CHALLENGE,
   COLLECTIBLE,